Move all serializers into a single file in the same dir
Update the serializer route on interface.backend.api
