+++ interface/backend/images/models.py
@@ -3,7 +3,6 @@ import glob
 import dicom
 from django.db import models
 from django.utils._os import safe_join
-from rest_framework import serializers
 
 
 class ImageSeries(models.Model):
@@ -38,12 +37,6 @@ class ImageSeries(models.Model):
             uri=uri)
 
 
-class ImageSeriesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ImageSeries
-        fields = ('id', 'patient_id', 'series_instance_uid', 'uri')
-
-
 class ImageLocation(models.Model):
     """
     Model representing a certain voxel location on certain image
@@ -55,11 +48,3 @@ class ImageLocation(models.Model):
     y = models.PositiveSmallIntegerField(help_text='Voxel index for Y axis, zero-index, from top left')
 
     z = models.PositiveSmallIntegerField(help_text='Slice index for Z axis, zero-index')
-
-
-class ImageLocationSerializer(serializers.ModelSerializer):
-    series = ImageSeriesSerializer()
-
-    class Meta:
-        model = ImageLocation
-        fields = ('id', 'series', 'x', 'y', 'z')
